Make the skip to contact text not show on hover
It will still show on tabbing to it (allowing users to skip tabbing through the nav menus) and screen readers will still read it. This should only cover the case where a user hovers, as they are using a mouse and the main content is already at the top of the screen, so the link can only be a point of confusion.